+++ src/types/read_file_part.rs
@@ -17,11 +17,11 @@ pub struct ReadFilePart {
     /// The offset from which to read the file
 
     #[serde(default)]
-    offset: i64,
+    offset: i32,
     /// Number of bytes to read. An error will be returned if there are not enough bytes available in the file from the specified position. Pass 0 to read all available data from the specified position
 
     #[serde(default)]
-    count: i64,
+    count: i32,
 
     #[serde(rename(serialize = "@type"))]
     td_type: String,
@@ -57,11 +57,11 @@ impl ReadFilePart {
         self.file_id
     }
 
-    pub fn offset(&self) -> i64 {
+    pub fn offset(&self) -> i32 {
         self.offset
     }
 
-    pub fn count(&self) -> i64 {
+    pub fn count(&self) -> i32 {
         self.count
     }
 }
@@ -81,12 +81,12 @@ impl RTDReadFilePartBuilder {
         self
     }
 
-    pub fn offset(&mut self, offset: i64) -> &mut Self {
+    pub fn offset(&mut self, offset: i32) -> &mut Self {
         self.inner.offset = offset;
         self
     }
 
-    pub fn count(&mut self, count: i64) -> &mut Self {
+    pub fn count(&mut self, count: i32) -> &mut Self {
         self.inner.count = count;
         self
     }
